VariableNode에 데이터 업데이트 및 처리 기능 개발 및 기초 테스트 완료
- NodeValidator 클래스에서 ord 속성이 null 일 때, fmt 속성 값에 오류가 없는 오류를 발견함.
- VariableNode 클래스에서는 이러한 경우 오류 로그를 남기지만 크래시는 발생하지 않도록 예외 처리 기능을 구현하였음
- 이때는 빈 문자열을 반환함

- 이제부터는 Node 설정에 대한 결정트리의 모든 path를 따라 최대한 다양한 테스트 케이스를 만들어서 검증해야 함

+++ lib/MUFFIN/src/Protocol/Modbus/Include/PolledData.h
@@ -35,8 +35,8 @@ namespace muffin { namespace modbus {
     public:
         Status UpdateCoil(const uint16_t address, const int8_t value);
         Status UpdateDiscreteInput(const uint16_t address, const int8_t value);
-        Status UpdateInputRegister(const uint16_t address, const uint16_t value);
-        Status UpdateHoldingRegister(const uint16_t address, const uint16_t value);
+        Status UpdateInputRegister(const uint16_t address, const int32_t value);
+        Status UpdateHoldingRegister(const uint16_t address, const int32_t value);
     public:
         datum_t RetrieveCoil(const uint16_t address) const;
         datum_t RetrieveDiscreteInput(const uint16_t address) const;
