+++ lib/MUFFIN/src/Protocol/Modbus/Include/PolledDataTable.h
@@ -33,8 +33,8 @@ namespace muffin { namespace modbus {
     public:
         Status UpdateCoil(const uint8_t slaveID, const uint16_t address, const int8_t value);
         Status UpdateDiscreteInput(const uint8_t slaveID, const uint16_t address, const int8_t value);
-        Status UpdateInputRegister(const uint8_t slaveID, const uint16_t address, const uint16_t value);
-        Status UpdateHoldingRegister(const uint8_t slaveID, const uint16_t address, const uint16_t value);
+        Status UpdateInputRegister(const uint8_t slaveID, const uint16_t address, const int32_t value);
+        Status UpdateHoldingRegister(const uint8_t slaveID, const uint16_t address, const int32_t value);
     public:
         datum_t RetrieveCoil(const uint8_t slaveID, const uint16_t address) const;
         datum_t RetrieveDiscreteInput(const uint8_t slaveID, const uint16_t address) const;
